feat(daemon): ephemeral self-signed QUIC cert when no cert configured
Reverse the persisted zero-config default: when neither `quic cert file`
nor `quic key file` is set, the QUIC listener mints a fresh self-signed
certificate in memory at bind time (as QuicAcceptor::bind already does)
rather than persisting one under a state dir. The identity is ephemeral -
never written to disk and regenerated every start.

resolve_quic_identity now returns a QuicIdentity enum: Files{cert,key}
verbatim when both directives are set, else Ephemeral. No disk writes, no
state directory, no rcgen dependency in the daemon crate.

Documented tradeoff (decision A): a TOFU quic_known_hosts client sees a
changed fingerprint each restart, so operators wanting a stable identity
must configure explicit cert/key files.
